Move Virtualenv out of `pex.tools.commands`.
