+++ tests/tools/commands/test_venv.py
@@ -22,9 +22,9 @@ from pex.interpreter import PythonInterpreter
 from pex.layout import Layout
 from pex.pex_builder import CopyMode, PEXBuilder
 from pex.testing import IS_PYPY, PY310, PY_VER, ensure_python_interpreter, run_pex_command
-from pex.tools.commands.virtualenv import Virtualenv
 from pex.typing import TYPE_CHECKING, cast
 from pex.util import named_temporary_file
+from pex.venv.virtualenv import Virtualenv
 
 if TYPE_CHECKING:
     from typing import Any, Dict, Iterable, Iterator, List, Optional, Protocol, Set, Text, Tuple
